pyproject.toml
[project]
name = "splitfree-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Rishav",email = "rishishah907@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "django-allauth (>=65.7.0,<66.0.0)",
    "django-cors-headers (>=4.3.1,<5.0.0)",
    "firebase-admin (>=6.7.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)",
    "google-genai (>=1.12.1,<2.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "splitfree_backend"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
